requirements.txt
docker>=3.0,<=4.0
GitPython==2.1.10
PyYAML==3.12
mypy==0.610
